pyproject.toml
[tool.poetry]
name = "terralab-cli"
# version is auto-incremented. do not edit.
version = "0.0.6"
description = "Command line interface for interacting with the Terra Scientific Pipelines Service, or Teaspoons."
authors = ["Terra Scientific Services <teaspoons-developers@broadinstitute.org>"]
readme = "README.md"
packages = [
    {include = "terralab"}
]

[tool.poetry.scripts]
terralab = "terralab.cli:cli"

[tool.poetry.dependencies]
python = "^3.9"
# FOR TESTING LOCALLY
# teaspoons-client = {path = "/Users/marymorg/workbench/terra-scientific-pipelines-service/python-client/generated", develop = true}
terra-scientific-pipelines-service-api-client = "^0.1.0"
python-dotenv = "^1.0.1"
click = "^8.1.7"
colorlog = "^6.9.0"
oauth2-cli-auth = "^1.5.0"
PyJWT = "^2.9.0"
requests = "^2.32.3"
tqdm = "^4.67.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^3.0.0"
mockito = ">=1.4.0"
pytest-mockito = ">=0.0.4"
ruff = "^0.7.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "terralab"

[tool.coverage.run]
branch = true
relative_files = true  # allows sonarcloud GHA step to process the coverage files

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "generated"
]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/test_*.py" = ["F811", "F401"]  # see https://github.com/astral-sh/ruff/issues/10662
